requirements.txt
opencv-python
numpy
Pillow
transformers
diffusers
accelerate
torch
